# backend/SanSo.Api.V6/ProgramCanonicalV8.cs
using System.Text;
using System.Text.Json.Serialization;
using Microsoft.AspNetCore.Routing;
using Npgsql;
using SanSo.Api;
using SanSo.Api.Modules;
using SanSo.Api.V2;
using SanSo.Api.V5;
using SanSo.Import;

var builder = WebApplication.CreateBuilder(args);
builder.Services.ConfigureHttpJsonOptions(options => options.SerializerOptions.Converters.Add(new JsonStringEnumConverter()));
var cs = builder.Configuration.GetConnectionString("Postgres") ?? Environment.GetEnvironmentVariable("SANSO_POSTGRES");
if (!builder.Environment.IsDevelopment() && string.IsNullOrWhiteSpace(cs))
    throw new InvalidOperationException("Production requires PostgreSQL");
if (cs is not null)
{
    builder.Services.AddSingleton(NpgsqlDataSource.Create(cs));
    builder.Services.AddSingleton<PostgresCommerceStore>();
    builder.Services.AddSingleton<PostgresLifecycleStore>();
    builder.Services.AddSingleton<PostgresImportStagingStoreV2>();
    builder.Services.AddSingleton<SanSo.Api.V6.PostgresNotificationStoreV1>();
}
builder.Services.AddSingleton<DemoStore>();
builder.Services.AddSanSoModules();
builder.Services.AddSingleton<IdentityService>();
builder.Services.AddSingleton<AuditTrail>();
builder.Services.AddSingleton<ReliableSync>();
builder.Services.AddSingleton<FinancialLifecycle>();
builder.Services.AddSingleton<TeamAndSupport>();
builder.Services.AddSingleton<BillingLifecycle>();
builder.Services.AddSingleton<SafeCopilot>();
builder.Services.AddCors(x => x.AddDefaultPolicy(p => p.WithOrigins("http://localhost:5173", "http://127.0.0.1:5174", "http://127.0.0.1:5176").AllowAnyHeader().AllowAnyMethod()));
var app = builder.Build();
app.UseMiddleware<SafeProblemMiddleware>();
app.UseCors();
IdentityEndpoints.MapIdentity(app);
V5AuthorizedComposition.MapSanSoModules(app);
var importer = new CommerceFileImporter(new ImportRegistry());
var memoryImport = new ImportConfirmationWorkflow();
var memoryNotifications = new NotificationCenter();

string Tenant(HttpRequest r) => r.HttpContext.Items["TenantId"] as string ?? r.Headers["X-Tenant-Id"].FirstOrDefault() ?? "tenant-an-nhien";
SessionPrincipal? Principal(HttpRequest r)
{
    if (r.HttpContext.Items["SessionPrincipal"] is SessionPrincipal p) return p;
    var tenant = r.Headers["X-Tenant-Id"].FirstOrDefault();
    var auth = r.Headers.Authorization.FirstOrDefault();
    if (tenant is null || auth?.StartsWith("Bearer ", StringComparison.OrdinalIgnoreCase) != true) return null;
    try { return app.Services.GetRequiredService<IdentityService>().Authenticate(auth[7..], tenant); }
    catch { return null; }
}
SessionPrincipal Require(HttpRequest r, string permission) => IdentityEndpoints.Require(r, app.Services.GetRequiredService<IdentityService>(), permission);
bool Db() => cs is not null;
PostgresCommerceStore Pg() => app.Services.GetRequiredService<PostgresCommerceStore>();
PostgresImportStagingStoreV2 Imports() => app.Services.GetRequiredService<PostgresImportStagingStoreV2>();
PostgresInventoryStoreV4 Inventory() => new(app.Services.GetRequiredService<NpgsqlDataSource>());
SanSo.Api.V6.PostgresNotificationStoreV1 NotificationStore() => app.Services.GetRequiredService<SanSo.Api.V6.PostgresNotificationStoreV1>();
void Prefer(RouteHandlerBuilder route) => route.Add(endpoint => ((RouteEndpointBuilder)endpoint).Order = -1);

Prefer(app.MapGet("/api/inventory/{sku}", async (HttpRequest r, string sku, CancellationToken ct) =>
{
    Require(r, "finance.read");
    return Db() ? Results.Ok(await Inventory().Get(Tenant(r), sku, ct)) : Results.Json(new { code = "DATABASE_REQUIRED" }, statusCode: 503);
}));
Prefer(app.MapPost("/api/inventory/{sku}/reserve", async (HttpRequest r, string sku, V7InventoryBody b, CancellationToken ct) =>
{
    Require(r, "inventory.write");
    if (!Db()) return Results.Json(new { code = "DATABASE_REQUIRED" }, statusCode: 503);
    try { return Results.Ok(await Inventory().Reserve(Tenant(r), sku, b.Quantity, b.SourceKey, ct)); }
    catch (InventoryConflictException) { return Results.Conflict(new { code = "INSUFFICIENT_ATP" }); }
}));
Prefer(app.MapPost("/api/inventory/{sku}/release", async (HttpRequest r, string sku, V7InventoryBody b, CancellationToken ct) =>
{
    Require(r, "inventory.write");
    if (!Db()) return Results.Json(new { code = "DATABASE_REQUIRED" }, statusCode: 503);
    try { return Results.Ok(await Inventory().Release(Tenant(r), sku, b.Quantity, b.SourceKey, ct)); }
    catch (InvalidOperationException e) { return Results.BadRequest(new { code = e.Message }); }
}));

Prefer(app.MapGet("/api/notifications", async (HttpRequest r, CancellationToken ct) =>
{
    Require(r, "finance.read");
    if (Db()) return Results.Ok(new { items = await NotificationStore().List(Tenant(r), ct), persisted = true });
    return Results.Ok(new { items = memoryNotifications.List(Tenant(r)), persisted = false });
}));
Prefer(app.MapPost("/api/notifications", async (HttpRequest r, V8NotificationBody b, CancellationToken ct) =>
{
    Require(r, "organization.manage");
    if (b.Channel != DeliveryChannel.InApp) return Results.Json(new { code = "EMAIL_PROVIDER_NOT_CONFIGURED" }, statusCode: 503);
    if (Db()) return Results.Ok(new { delivery = await NotificationStore().RaiseInApp(Tenant(r), b.Type, b.ResourceRef, b.WindowStart, ct), persisted = true });
    return Results.Ok(new { delivery = memoryNotifications.Raise(Tenant(r), b.Type, b.Channel, b.Recipient, b.ResourceRef, b.WindowStart), persisted = false });
}));
Prefer(app.MapPost("/api/notifications/{id}/acknowledge", async (HttpRequest r, string id, CancellationToken ct) =>
{
    Require(r, "finance.read");
    try
    {
        if (Db()) return Results.Ok(new { delivery = await NotificationStore().Acknowledge(Tenant(r), id, ct), persisted = true });
        return Results.Ok(new { delivery = memoryNotifications.Acknowledge(Tenant(r), id), persisted = false });
    }
    catch (KeyNotFoundException) { return Results.NotFound(new { code = "NOTIFICATION_NOT_FOUND" }); }
}));
app.MapGet("/health", async (CancellationToken ct) =>
{
    if (!Db()) return Results.Ok(new { status = "degraded", database = "not-configured", api = "v7" });
    try
    {
        await using var c = await app.Services.GetRequiredService<NpgsqlDataSource>().OpenConnectionAsync(ct);
        await using var q = c.CreateCommand();
        q.CommandText = "SELECT 1";
        await q.ExecuteScalarAsync(ct);
        return Results.Ok(new { status = "ok", database = "postgresql", api = "v7" });
    }
    catch { return Results.Json(new { status = "unhealthy" }, statusCode: 503); }
});

app.MapPost("/api/imports/preview", async (HttpRequest r, CancellationToken ct) =>
{
    if (!r.HasFormContentType) return Results.BadRequest(new { code = "MULTIPART_REQUIRED" });
    var file = (await r.ReadFormAsync(ct)).Files.GetFile("file");
    if (file is null) return Results.BadRequest(new { code = "FILE_REQUIRED" });
    if (file.Length > 10 * 1024 * 1024) return Results.BadRequest(new { code = "FILE_TOO_LARGE" });
    await using var stream = file.OpenReadStream();
    ImportPreview preview;
    try
    {
        var ext = Path.GetExtension(file.FileName).ToLowerInvariant();
        if (ext == ".csv" && file.ContentType is "text/csv" or "application/csv" or "application/vnd.ms-excel" or "application/octet-stream")
            preview = importer.PreviewCsv(Tenant(r), stream);
        else if (ext == ".xlsx" && file.ContentType is "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet" or "application/octet-stream")
            preview = importer.PreviewXlsx(Tenant(r), stream);
        else return Results.BadRequest(new { code = "UNSUPPORTED_FILE_TYPE" });
    }
    catch (DecoderFallbackException) { return Results.BadRequest(new { code = "ENCODING_INVALID", hint = "Save CSV as UTF-8" }); }
    catch (InvalidDataException e) { return Results.BadRequest(new { code = e.Message }); }
    if (Db())
    {
        var stored = await Imports().Stage(Tenant(r), preview, ct);
        return Results.Ok(new { stored.BatchId, stored.PreviewToken, stored.ExpiresAt, preview.Format, preview.TemplateVersion, preview.Checksum, preview.Delimiter, preview.Headers, preview.Rows, preview.Errors, preview.Duplicate, persisted = true });
    }
    var staged = memoryImport.Stage(Tenant(r), preview);
    return Results.Ok(new { batchId = (string?)null, staged.PreviewToken, staged.ExpiresAt, preview.Format, preview.TemplateVersion, preview.Checksum, preview.Delimiter, preview.Headers, preview.Rows, preview.Errors, preview.Duplicate, persisted = false });
});

app.MapPost("/api/imports/confirm", async (HttpRequest r, V7ConfirmBody b, CancellationToken ct) =>
{
    if (Db())
    {
        var actor = Principal(r)?.UserId;
        if (actor is null) return Results.Unauthorized();
        try
        {
            var x = await Imports().Confirm(Tenant(r), b.PreviewToken, b.Checksum, actor, ct);
            return Results.Ok(new { x.BatchId, x.Checksum, x.AcceptedRows, x.RejectedRows, x.Duplicate, persisted = true });
        }
        catch (InvalidOperationException e) { return Results.BadRequest(new { code = e.Message }); }
    }
    var m = memoryImport.Confirm(Tenant(r), b.PreviewToken, b.Checksum);
    return Results.Ok(new { m.Checksum, m.AcceptedRows, m.RejectedRows, m.Duplicate, persisted = false });
});
app.MapGet("/api/dashboard", async (HttpRequest r, DemoStore demo, CancellationToken ct) => Db() ? Results.Ok(await Pg().Dashboard(Tenant(r), ct)) : Results.Ok(demo.Dashboard(Tenant(r))));
app.MapGet("/api/orders", async (HttpRequest r, DemoStore demo, CancellationToken ct) => Db() ? Results.Ok(await Pg().Orders(Tenant(r), ct)) : Results.Ok(demo.Orders(Tenant(r))));
app.MapGet("/api/reconciliations/current", async (HttpRequest r, DemoStore demo, CancellationToken ct) => Db() ? Results.Ok(await Pg().CurrentReconciliation(Tenant(r), ct)) : Results.Ok(demo.Reconcile(Tenant(r))));
app.Run();

public partial class Program { }
record V7InventoryBody(int Quantity, string SourceKey);
record V7ConfirmBody(string PreviewToken, string Checksum);
record V8NotificationBody(NotificationType Type, DeliveryChannel Channel, string Recipient, string ResourceRef, DateTimeOffset WindowStart);