# docs/openapi-v14-canonical.json
{
    "openapi":  "3.1.0",
    "info":  {
                 "title":  "SanSo API V14",
                 "version":  "0.14.0-pilot",
                 "description":  "Canonical V14 pilot contract. Adds formula-free reconciliation XLSX data packs to the durable preview/checksum/confirm/download lifecycle."
             },
    "servers":  [
                    {
                        "url":  "http://localhost:5080"
                    }
                ],
    "security":  [
                     {
                         "bearerAuth":  [

                                        ]
                     }
                 ],
    "paths":  {
...
                  "/api/imports/preview":  {
                                               "post":  {
                                                            "parameters":  [
                                                                               {
                                                                                   "$ref":  "#/components/parameters/TenantHeader"
                                                                               }
                                                                           ],
                                                            "requestBody":  {
                                                                                "required":  true,
                                                                                "content":  {
                                                                                                "multipart/form-data":  {
                                                                                                                            "schema":  {
                                                                                                                                           "type":  "object",
                                                                                                                                           "required":  [
                                                                                                                                                            "file"
                                                                                                                                                        ],
                                                                                                                                           "properties":  {
                                                                                                                                                              "file":  {
                                                                                                                                                                           "type":  "string",
                                                                                                                                                                           "format":  "binary",
                                                                                                                                                                           "maxLength":  10485760,
                                                                                                                                                                           "description":  "CSV or XLSX allowlisted by extension and MIME; maximum 10 MiB; CSV must be UTF-8."
                                                                                                                                                                       }
                                                                                                                                                          }
                                                                                                                                       }
                                                                                                                        }
                                                                                            }
                                                                            },
                                                            "responses":  {
                                                                              "200":  {
                                                                                          "description":  "Preview only; no business projection written",
                                                                                          "content":  {
                                                                                                          "application/json":  {
                                                                                                                                   "schema":  {
                                                                                                                                                  "$ref":  "#/components/schemas/ImportPreview"
                                                                                                                                              }
                                                                                                                               }
                                                                                                      }
                                                                                      },
                                                                              "401":  {
                                                                                          "$ref":  "#/components/responses/Unauthorized"
                                                                                      },
                                                                              "400":  {
                                                                                          "description":  "Safe import validation error"
                                                                                      }
                                                                          }
                                                        }
                                           },
                  "/api/imports/confirm":  {
                                               "post":  {
                                                            "parameters":  [
                                                                               {
                                                                                   "$ref":  "#/components/parameters/TenantHeader"
                                                                               }
                                                                           ],
                                                            "requestBody":  {
                                                                                "required":  true,
                                                                                "content":  {
                                                                                                "application/json":  {
                                                                                                                         "schema":  {
                                                                                                                                        "$ref":  "#/components/schemas/ImportConfirmRequest"
                                                                                                                                    }
                                                                                                                     }
                                                                                            }
                                                                            },
                                                            "responses":  {
                                                                              "200":  {
                                                                                          "description":  "Explicit confirmation result",
                                                                                          "content":  {
                                                                                                          "application/json":  {
                                                                                                                                   "schema":  {
                                                                                                                                                  "$ref":  "#/components/schemas/ImportConfirmation"
                                                                                                                                              }
                                                                                                                               }
                                                                                                      }
                                                                                      },
                                                                              "400":  {
                                                                                          "$ref":  "#/components/responses/BadRequest"
                                                                                      },
                                                                              "401":  {
                                                                                          "$ref":  "#/components/responses/Unauthorized"
                                                                                      }
                                                                          }
                                                        }
                                           },
                  "/api/dashboard":  {
                                         "get":  {
                                                     "parameters":  [
                                                                        {
                                                                            "$ref":  "#/components/parameters/TenantHeader"
                                                                        }
                                                                    ],
                                                     "responses":  {
                                                                       "200":  {
                                                                                   "description":  "Tenant dashboard"
                                                                               }
                                                                   }
                                                 }
                                     },
                  "/api/orders":  {
                                      "get":  {
                                                  "parameters":  [
                                                                     {
                                                                         "$ref":  "#/components/parameters/TenantHeader"
                                                                     }
                                                                 ],
                                                  "responses":  {
                                                                    "200":  {
                                                                                "description":  "Tenant orders"
                                                                            }
                                                                }
                                              }
                                  },
                  "/api/reconciliations/current":  {
                                                       "get":  {
                                                                   "parameters":  [
                                                                                      {
                                                                                          "$ref":  "#/components/parameters/TenantHeader"
                                                                                      }
                                                                                  ],
                                                                   "responses":  {
                                                                                     "200":  {
                                                                                                 "description":  "Line-level reconciliation and explanations"
                                                                                             }
                                                                                 }
                                                               }
                                                   },
...
                  "/api/inventory/{sku}/reserve":  {
                                                       "post":  {
                                                                    "parameters":  [
                                                                                       {
                                                                                           "$ref":  "#/components/parameters/TenantHeader"
                                                                                       },
                                                                                       {
                                                                                           "name":  "sku",
                                                                                           "in":  "path",
                                                                                           "required":  true,
                                                                                           "schema":  {
                                                                                                          "type":  "string"
                                                                                                      }
                                                                                       }
                                                                                   ],
                                                                    "responses":  {
                                                                                      "409":  {
                                                                                                  "description":  "Insufficient ATP"
                                                                                              },
                                                                                      "200":  {
                                                                                                  "content":  {
                                                                                                                  "application/json":  {
                                                                                                                                           "schema":  {
                                                                                                                                                          "$ref":  "#/components/schemas/InventorySnapshot"
                                                                                                                                                      }
                                                                                                                                       }
                                                                                                              },
                                                                                                  "description":  "Durable inventory snapshot"
                                                                                              },
                                                                                      "401":  {
                                                                                                  "$ref":  "#/components/responses/Unauthorized"
                                                                                              },
                                                                                      "403":  {
                                                                                                  "$ref":  "#/components/responses/Forbidden"
                                                                                              },
                                                                                      "503":  {
                                                                                                  "$ref":  "#/components/responses/ServiceUnavailable"
                                                                                              }
                                                                                  },
                                                                    "requestBody":  {
                                                                                        "content":  {
                                                                                                        "application/json":  {
                                                                                                                                 "schema":  {
                                                                                                                                                "$ref":  "#/components/schemas/InventoryMutationRequest"
                                                                                                                                            }
                                                                                                                             }
                                                                                                    },
                                                                                        "required":  true
                                                                                    }
                                                                }
                                                   },
...
                  "/api/notifications":  {
                                             "get":  {
                                                         "parameters":  [
                                                                            {
                                                                                "$ref":  "#/components/parameters/TenantHeader"
                                                                            }
                                                                        ],
                                                         "responses":  {
                                                                           "200":  {
                                                                                       "description":  "Tenant inbox with masked recipient"
                                                                                   }
                                                                       }
                                                     },
                                             "post":  {
                                                          "parameters":  [
                                                                             {
                                                                                 "$ref":  "#/components/parameters/TenantHeader"
                                                                             }
                                                                         ],
                                                          "responses":  {
                                                                            "200":  {
                                                                                        "description":  "Development envelope with persisted=false"
                                                                                    },
                                                                            "503":  {
                                                                                        "description":  "Email provider or production dependency is not configured"
                                                                                    }
                                                                        },
                                                          "requestBody":  {
                                                                              "content":  {
                                                                                              "application/json":  {
                                                                                                                       "schema":  {
                                                                                                                                      "$ref":  "#/components/schemas/NotificationRequest"
                                                                                                                                  }
                                                                                                                   }
                                                                                          },
                                                                              "required":  true
                                                                          }
                                                      }
                                         },
...
              },
    "components":  {
                       "securitySchemes":  {
                                               "bearerAuth":  {
                                                                  "type":  "http",
                                                                  "scheme":  "bearer",
                                                                  "bearerFormat":  "opaque"
                                                              }
                                           },
                       "parameters":  {
                                          "TenantHeader":  {
                                                               "name":  "X-Tenant-Id",
                                                               "in":  "header",
                                                               "required":  true,
                                                               "schema":  {
                                                                              "type":  "string",
                                                                              "minLength":  1
                                                                          }
                                                           },
...
                                      },
                       "responses":  {
                                         "BadRequest":  {
                                                            "description":  "Safe validation response"
                                                        },
                                         "Unauthorized":  {
                                                              "description":  "Missing, invalid, expired or tenant-mismatched session"
                                                          },
                                         "Forbidden":  {
                                                           "description":  "Authenticated role lacks permission"
                                                       },
                                         "ServiceUnavailable":  {
                                                                    "description":  "Required dependency unavailable"
                                                                }
                                     },
...
                   },
...
}